requirements.txt
appdirs==1.4.3
asgiref==3.3.1
CacheControl==0.12.6
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
colorama==0.4.3
contextlib2==0.6.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.4.4
defusedxml==0.6.0
distlib==0.3.0
distro==1.4.0
dj-rest-auth==2.1.3
Django==3.1.6
django-allauth==0.44.0
django-rest-swagger==2.2.0
django-templated-mail==1.1.1
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
djoser==2.1.0
drf-yasg==1.20.0
gunicorn==20.0.4
html5lib==1.0.1
idna==2.10
inflection==0.5.1
ipaddr==2.2.0
itypes==1.2.0
Jinja2==2.11.3
lockfile==0.12.2
MarkupSafe==1.1.1
msgpack==0.6.2
oauthlib==3.1.0
openapi-codec==1.3.2
packaging==20.3
pep517==0.8.2
Pillow==8.1.0
progress==1.5
pycparser==2.20
PyJWT==2.0.1
pyparsing==2.4.6
python3-openid==3.2.0
pytoml==0.1.21
pytz==2021.1
requests==2.25.1
requests-oauthlib==1.3.0
retrying==1.3.3
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2
simplejson==3.17.2
six==1.15.0
social-auth-app-django==4.0.0
social-auth-core==4.0.3
sqlparse==0.4.1
uritemplate==3.0.1
urllib3==1.26.3
webencodings==0.5.1
